Backend/requirements.txt
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.0.0
matplotlib==3.7.2
scikit-learn==1.3.0
transformers==4.35.0
torch==2.1.0
requests==2.31.0
opencv-python==4.8.1.78